ServerSentEvents/submissions/MOSQUITO/field-name-event.html

<!DOCTYPE html>
<html>
<head>
	<title>The steps to process the field given a field name and a field value depend on the field name: The field name is "event":</title>
  <meta charset="utf-8" />
  	<meta name=viewport content="width=device-width, maximum-scale=1.0, user-scalable=no" />
	  <link rel="author" title="Fabrice Clari" href="mailto:f.clari@inno-group.com">
	  <link rel="author" title="Dimitri Bocquet" href="mailto:Dimitri.Bocquet@mosquito-fp7.eu">
	<script src="/resources/testharness.js"></script>
	<script src="/resources/testharnessreport.js"></script>
	<script src="/resources/WebIDLParser.js"></script>
	<script src="/resources/idlharness.js"></script>
</head>
<body>
	<h1>The steps to process the field given a field name and a field value depend on the field name: The field name is "event":</h1>
	<div id="log"></div>
  <script>
  	setup({timeout:1500});
    var s = new EventSource('http://html5.mosquito-fp7.eu/messages/t0074');
    vt = null;
    t = async_test("field name is 'event': set the event name buffer to field value.");
    s.onopen = function(e){ 
    	/* Connection opened. */
    };
    	
    s.addEventListener("testevent", function(e){
    	t.step(function(){assert_equals(e.type, "testevent")});
    	t.done();
    }, false);
    
    //	console.log(s.testevent);
    s.onerror = function(e){
    	/* Connection closed. */
    };
    

  </script>
  </div>
</body>
</html>
